Create README.md
A company in Chile markets sensors to retail stores so that the stores can better keep track of the traffic in the stores, clients who made purchases, number of salespeople in the stores, etc. The data MiniProj3.csv has daily data for 30 stores in Chile over a year. First column is code of store, second column is name of store which is also city where the store is, then date of sensor data as given by year, month, day, name of day (in Spanish), week of the year (a number between 1 and 52), traffic data (0 if sensor was malfunctioning or not installed yet), number of customer, sales, average ticket (=average amount on receipt), number of salespeople who worked that day, sales per salesperson, number of tickets (receipts) for salesperson, average number of products per ticket (receipt), average price in Chilean pesos, average number of salespeople working at the same time. You do not have to use all the data for your project. 0 in the data means missing data (sensor didn’t detect anything, or was not installed yet) or the store was closed. You can use R or Excel as software for the plots and summary tables (PivotTable in Excel). Hierarchical clustering and k-means clustering must be done in R, obviously.